+++ src/test/java/com/openggf/sprites/playable/TestSidekickCpuDespawnParity.java
@@ -2,8 +2,6 @@ package com.openggf.sprites.playable;
 
 import com.openggf.tests.TestEnvironment;
 import com.openggf.game.session.SessionManager;
-import com.openggf.game.session.EngineServices;
-import com.openggf.game.session.EngineContext;
 import com.openggf.game.GameServices;
 import com.openggf.game.PhysicsFeatureSet;
 import com.openggf.game.PlayableEntity;
@@ -24,8 +22,7 @@ class TestSidekickCpuDespawnParity {
 
     @BeforeEach
     void configureRuntime() {
-        EngineServices.configure(EngineContext.fromLegacySingletonsForBootstrap());
-        TestEnvironment.activeGameplayMode();
+        TestEnvironment.resetAll();
     }
 
     @AfterEach